--- fixup.sh
@@ -1,10 +1,0 @@
-cat >dist/cjs/package.json <<!EOF
-{
-    "type": "commonjs"
-}
-!EOF
-cat >dist/mjs/package.json <<!EOF
-{
-    "type": "module"
-}
-!EOF
