+++ package.json
@@ -3,16 +3,12 @@
   "version": "3.0.0",
   "description": "Instances of classes which are directly callable as functions.",
   "repository": "CGamesPlay/node-callable-instance",
-  "main": "./dist/cjs/index.js",
-  "module": "./dist/mjs/index.js",
-  "types": "lib/index.d.ts",
-  "exports": {
-    "import": "./dist/mjs/index.js",
-    "require": "./dist/cjs/index.js"
-  },
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "type": "commonjs",
   "scripts": {
-    "build": "rm -rf dist && tsc && tsc -p tsconfig-cjs.json && ./fixup.sh && npm link && npm link callable-instance",
-    "test": "npm run build && mocha src/test"
+    "link": "npm link && npm link callable-instance",
+    "test": "npm run link && mocha test"
   },
   "keywords": [
     "instance",
