+++ test/classGeneric.ts
