+++ test/funcGeneric.ts
