+++ test/interfaceGeneric.ts
