--- tsconfig-cjs.json
@@ -1,8 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "CommonJS",
-        "outDir": "dist/cjs",
-        "target": "es6",
-    },
-}
