--- tsconfig.json
@@ -1,35 +1,0 @@
-{
-    "compilerOptions": {
-        /// for cjs - tsconfig-cjs.json ///
-        "module": "es6",
-        "outDir": "dist/mjs",
-        "target": "es6",
-        /// for cjs - tsconfig-cjs.json ///
-        "removeComments": true,
-        "allowJs": true,
-        "allowSyntheticDefaultImports": true,
-        "baseUrl": "src",
-        "declaration": false,
-        "esModuleInterop": true,
-        "inlineSourceMap": false,
-        "listEmittedFiles": false,
-        "listFiles": false,
-        "moduleResolution": "node",
-        "pretty": true,
-        "rootDir": "src",
-        "skipLibCheck": true,
-        "strict": true,
-        "traceResolution": false,
-        "typeRoots": [
-            "node_modules/@types", "lib"
-        ]
-    },
-    "include": [ "src" ],
-    "exclude": [
-        "test",
-        "src/types",
-        "src/test",
-        "lib",
-        "node_modules",
-    ]
-}
